third_party/blink/web_tests/platform/win11-arm64/external/wpt/speech-api/SpeechRecognition-onerror.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: expected "network" but got "audio-capture"
[TIMEOUT] Speech recognition onerror event is called.
  Test timed out
Harness: the test ran to completion.
